+++ class-components/src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
-import ErrorBoundary from './components/error boundery/error-boundery.tsx'
+import ErrorBoundary from './components/errorBoundery/error-boundery.tsx'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
